requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
pytz==2020.5
sqlparse==0.4.1
